Parameters and datasets

+++ include/detectors/Bold3DMDetector.h
@@ -16,7 +16,7 @@ namespace visy {
 
         class Bold3DMDetector : public Detector {
         public:
-            Bold3DMDetector(std::vector<int>& sizes, int n_bins = 12, bool filter_occlusion = true, float zone_radius = 5.0f, float zone_slice = 2.0f, float area_normals_angular_th = 25.0f, float area_max_distance = 0.001f);
+            Bold3DMDetector(std::vector<float>& sizes, int n_bins = 12, bool filter_occlusion = true, float zone_radius = 5.0f, float zone_slice = 2.0f, float area_normals_angular_th = 25.0f, float area_max_distance = 0.001f);
             virtual void detect(cv::Mat& source, pcl::PointCloud<PointType>::Ptr cloud, std::vector<visy::extractors::KeyPoint3D>& keypoints, cv::Mat& descriptor,cv::Mat* mask = NULL);
             virtual ~Bold3DMDetector();
             virtual std::string buildNameImpl();
